Modularise non-category routes
